# core-lending-loan-origination-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-loan-origination
  description: Loan Origination Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/loan-origination
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/loan-applications/{applicationId}/scores/{scoreId}:
    get:
      tags:
        - UnderwritingScore
      summary: Get a score
      description: Fetch a specific underwriting score record by ID.
      operationId: getScore
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scoreId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UnderwritingScoreDTO'
    put:
      tags:
        - UnderwritingScore
      summary: Update a score
      description: Updates an existing underwriting score record.
      operationId: updateScore
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scoreId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnderwritingScoreDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UnderwritingScoreDTO'
    delete:
      tags:
        - UnderwritingScore
      summary: Delete a score
      description: Removes a scoring record from the application.
      operationId: deleteScore
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scoreId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/loan-applications/{applicationId}/scores:
    get:
      tags:
        - UnderwritingScore
      summary: List scores
      description: Retrieves a paginated list of underwriting scores for a loan application.
      operationId: findAllScores
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pageNumber
          in: query
          description: The zero-based page number to retrieve.
          required: false
          schema:
            type: string
            description: The zero-based page number to retrieve.
            example: 0
            default: '0'
          example: 0
        - name: pageSize
          in: query
          description: The number of items per page.
          required: false
          schema:
            type: string
            description: The number of items per page.
            example: 10
            default: '10'
          example: 10
        - name: sortBy
          in: query
          description: The field to sort the results by.
          required: false
          schema:
            type: string
            description: The field to sort the results by.
            example: name
          example: name
        - name: sortDirection
          in: query
          description: The direction of sorting, either ASC or DESC.
          required: false
          schema:
            type: string
            description: The direction of sorting, either ASC or DESC.
            example: DESC
            default: DESC
          example: DESC
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseUnderwritingScoreDTO'
    post:
      tags:
        - UnderwritingScore
      summary: Create a new score
      description: Adds a new underwriting score record.
      operationId: createScore
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnderwritingScoreDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UnderwritingScoreDTO'
...
components:
  schemas:
...
    UnderwritingScoreDTO:
      required:
        - loanApplicationId
        - scoreValue
        - scorecardName
        - scoringTimestamp
      type: object
      properties:
        underwritingScoreId:
          type: string
          format: uuid
          readOnly: true
        loanApplicationId:
          type: string
          format: uuid
        scoreValue:
          maximum: 1000
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        scorecardName:
          maxLength: 100
          minLength: 0
          type: string
        reasonCodes:
          maxLength: 2000
          minLength: 0
          type: string
        scoringTimestamp:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationResponseUnderwritingScoreDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/UnderwritingScoreDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...